Add Cloud Support to IntelliFire integration. This goes in turn with a large refactor of backing library based on discussion with HTT (Home and Hearth Technologies) who make the units.

+++ homeassistant/components/intellifire/climate.py
@@ -63,7 +63,7 @@ class IntellifireClimate(IntellifireEntity, ClimateEntity):
         super().__init__(coordinator, description)
 
         if coordinator.data.thermostat_on:
-            self.last_temp = coordinator.data.thermostat_setpoint_c
+            self.last_temp = (int)(coordinator.data.thermostat_setpoint_c)
 
     @property
     def hvac_mode(self) -> HVACMode:
